invertedColourRandomLines
First project. Generates lines with random lengths with one endpoint in
the center of the window and the other at a random location. Clicking
the mouse leads to: generating a new, random number of lines; giving all
of those lines a random colour; changing the background colour to be the
inverse of the line colour; and changing the framerate to a random
number.